Starting repository development

+++ src/main/java/com/itesm/panoptimize/repository/AgentPerformanceRepository.java
@@ -4,7 +4,10 @@ import com.itesm.panoptimize.model.AgentPerformance;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+/**
+ * Repository interface for {@link AgentPerformance} instances.
+ */
 @Repository
-public interface AgentPerformanceRepository extends JpaRepository<AgentPerformance, Long> {
+public interface AgentPerformanceRepository extends JpaRepository<AgentPerformance, Integer> {
 
 }
