+++ src/main/java/com/itesm/panoptimize/repository/CompanyRepository.java
@@ -4,6 +4,9 @@ import com.itesm.panoptimize.model.Company;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+/**
+ * Repository for {@link Company} instances.
+ */
 @Repository
-public interface CompanyRepository extends JpaRepository<Company, Long> {
+public interface CompanyRepository extends JpaRepository<Company, Integer> {
 }
